Remove dataUtils
